Issue 837 async (#132)
* began work on socket server refactor

* async refactor

* mongo checks for processor status
