+++ unfetter-socket-server/src/models/mongoose/config.ts
@@ -1,0 +1,4 @@
+import { Schema, Model, model } from 'mongoose';
+
+const configModel: Model<any> = model<any>('config', new Schema({}, { strict: false }), 'config');
+export default configModel;
