+++ unfetter-socket-server/src/models/mongoose/stix.ts
@@ -1,0 +1,4 @@
+import { Schema, Model, model } from 'mongoose';
+
+const stixModel: Model<any> = model<any>('stix', new Schema({}, { strict: false }), 'stix');
+export default stixModel;
