+++ unfetter-socket-server/src/server/app.ts
@@ -6,7 +6,6 @@ import * as publishController from '../controllers/publish';
 
 const app: any = express();
 
-// app.use(cors());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
@@ -20,10 +19,7 @@ app.get('/heartbeat', (req: express.Request, res: express.Response) => {
 app.use((req: express.Request, res: express.Response, next: any) => {
     let err = new Error('Not Found');
     console.log('catching 404 error');
-    // console.log(req.url);
     return res.status(404).json({'error': '404 not found'});
-    // res.status(404);
-    // return next(err);
 });
 
 export default app;
