+++ unfetter-socket-server/src/vendor.d.ts
@@ -1,5 +1,7 @@
 declare module NodeJS {
     interface Global {
-        conn: any
+        conn: any,
+        unfetterconfigurations: any,
+        unfetteropenid: string
     }
 }
